made values annotatable   - implemented clone method for all types of values   - added annotation methods to IValue   - added implementation of annotation methods to abstract Value class in hash implementation
Note that since values are immutable, setting an annotation triggers a clone of the value, this clone is quite cheap though because the actual object references helt by the value do not have to be cloned, since they are immutable by themselves. Care must be taken for values that are still in a writeable state though.